CB-12902 Add AwsNativeSetup. The old AwsSetup splitted 2 setup implementations, one for the cloudformation and one for the common. The Native way doesn't need to implement the scaling prerequisitet
