+++ cloud-aws-common/src/main/java/com/sequenceiq/cloudbreak/cloud/aws/common/AwsSetup.java
@@ -1,9 +1,8 @@
-package com.sequenceiq.cloudbreak.cloud.aws;
+package com.sequenceiq.cloudbreak.cloud.aws.common;
 
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
-import java.util.stream.Collectors;
 
 import javax.inject.Inject;
 
@@ -11,14 +10,9 @@ import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Component;
 
 import com.amazonaws.AmazonClientException;
 import com.amazonaws.AmazonServiceException;
-import com.amazonaws.services.autoscaling.model.AutoScalingGroup;
-import com.amazonaws.services.autoscaling.model.DescribeAutoScalingGroupsRequest;
-import com.amazonaws.services.autoscaling.model.DescribeAutoScalingGroupsResult;
-import com.amazonaws.services.autoscaling.model.Instance;
 import com.amazonaws.services.ec2.model.DescribeInternetGatewaysRequest;
 import com.amazonaws.services.ec2.model.DescribeInternetGatewaysResult;
 import com.amazonaws.services.ec2.model.DescribeKeyPairsRequest;
@@ -29,27 +23,20 @@ import com.amazonaws.services.ec2.model.InternetGateway;
 import com.amazonaws.services.ec2.model.InternetGatewayAttachment;
 import com.amazonaws.services.ec2.model.Subnet;
 import com.sequenceiq.cloudbreak.cloud.Setup;
-import com.sequenceiq.cloudbreak.cloud.aws.client.AmazonAutoScalingClient;
-import com.sequenceiq.cloudbreak.cloud.aws.client.AmazonCloudFormationClient;
-import com.sequenceiq.cloudbreak.cloud.aws.common.AwsInstanceConnector;
-import com.sequenceiq.cloudbreak.cloud.aws.common.AwsPlatformResources;
 import com.sequenceiq.cloudbreak.cloud.aws.common.client.AmazonEc2Client;
 import com.sequenceiq.cloudbreak.cloud.aws.common.view.AuthenticatedContextView;
 import com.sequenceiq.cloudbreak.cloud.aws.common.view.AwsCredentialView;
-import com.sequenceiq.cloudbreak.cloud.aws.view.AwsInstanceView;
-import com.sequenceiq.cloudbreak.cloud.aws.view.AwsNetworkView;
+import com.sequenceiq.cloudbreak.cloud.aws.common.view.AwsInstanceView;
+import com.sequenceiq.cloudbreak.cloud.aws.common.view.AwsNetworkView;
 import com.sequenceiq.cloudbreak.cloud.context.AuthenticatedContext;
 import com.sequenceiq.cloudbreak.cloud.exception.CloudConnectorException;
 import com.sequenceiq.cloudbreak.cloud.model.AvailabilityZone;
 import com.sequenceiq.cloudbreak.cloud.model.CloudCredential;
-import com.sequenceiq.cloudbreak.cloud.model.CloudInstance;
 import com.sequenceiq.cloudbreak.cloud.model.CloudRegions;
 import com.sequenceiq.cloudbreak.cloud.model.CloudStack;
-import com.sequenceiq.cloudbreak.cloud.model.CloudVmInstanceStatus;
 import com.sequenceiq.cloudbreak.cloud.model.Group;
 import com.sequenceiq.cloudbreak.cloud.model.Image;
 import com.sequenceiq.cloudbreak.cloud.model.InstanceAuthentication;
-import com.sequenceiq.cloudbreak.cloud.model.InstanceStatus;
 import com.sequenceiq.cloudbreak.cloud.model.Location;
 import com.sequenceiq.cloudbreak.cloud.model.SpiFileSystem;
 import com.sequenceiq.cloudbreak.cloud.notification.PersistenceNotifier;
@@ -57,8 +44,7 @@ import com.sequenceiq.cloudbreak.util.DocumentationLinkProvider;
 import com.sequenceiq.common.api.type.ImageStatus;
 import com.sequenceiq.common.api.type.ImageStatusResult;
 
-@Component
-public class AwsSetup implements Setup {
+public abstract class AwsSetup implements Setup {
     private static final Logger LOGGER = LoggerFactory.getLogger(AwsSetup.class);
 
     private static final String IGW_DOES_NOT_EXIST_MSG = "The given internet gateway '%s' does not exist or belongs to a different region.";
@@ -84,16 +70,10 @@ public class AwsSetup implements Setup {
     private boolean awsSpotinstanceEnabled;
 
     @Inject
-    private CloudFormationStackUtil cfStackUtil;
-
-    @Inject
-    private LegacyAwsClient awsClient;
+    private CommonAwsClient awsClient;
 
     @Inject
     private AwsPlatformResources awsPlatformResources;
-
-    @Inject
-    private AwsInstanceConnector instanceConnector;
 
     @Override
     public ImageStatusResult checkImageStatus(AuthenticatedContext authenticatedContext, CloudStack stack, Image image) {
@@ -223,48 +203,4 @@ public class AwsSetup implements Setup {
         }
     }
 
-    @Override
-    public void scalingPrerequisites(AuthenticatedContext ac, CloudStack stack, boolean upscale) {
-        if (!upscale) {
-            return;
-        }
-        String regionName = ac.getCloudContext().getLocation().getRegion().value();
-        AwsCredentialView awsCredential = new AwsCredentialView(ac.getCloudCredential());
-        AmazonCloudFormationClient cloudFormationClient = awsClient.createCloudFormationClient(awsCredential, regionName);
-        AmazonAutoScalingClient amazonASClient = awsClient.createAutoScalingClient(awsCredential, regionName);
-        List<Group> groups = stack.getGroups().stream().filter(g -> g.getInstances().stream().anyMatch(
-                inst -> InstanceStatus.CREATE_REQUESTED == inst.getTemplate().getStatus())).collect(Collectors.toList());
-        Map<String, Group> groupMap = groups.stream().collect(
-                Collectors.toMap(g -> cfStackUtil.getAutoscalingGroupName(ac, cloudFormationClient, g.getName()), g -> g));
-        DescribeAutoScalingGroupsResult result = amazonASClient.describeAutoScalingGroups(
-                new DescribeAutoScalingGroupsRequest().withAutoScalingGroupNames(groupMap.keySet()));
-        for (AutoScalingGroup asg : result.getAutoScalingGroups()) {
-            Group group = groupMap.get(asg.getAutoScalingGroupName());
-            List<CloudInstance> groupInstances = group.getInstances().stream().filter(
-                    inst -> InstanceStatus.CREATED.equals(inst.getTemplate().getStatus())).collect(Collectors.toList());
-            List<CloudVmInstanceStatus> instanceStatuses = instanceConnector.check(ac, groupInstances);
-            if (!instanceStatuses.stream().allMatch(inst -> inst.getStatus().equals(InstanceStatus.STARTED))) {
-                String errorMessage = "Not all the existing instances are in [Started] state, upscale is not possible!";
-                LOGGER.info(errorMessage);
-                throw new CloudConnectorException(errorMessage);
-            }
-            List<Instance> asgOnlyInstances = asg.getInstances().stream()
-                    .filter(inst -> groupInstances.stream().noneMatch(gi -> gi.getInstanceId().equals(inst.getInstanceId()))).collect(Collectors.toList());
-            List<CloudInstance> cbOnlyInstances = groupInstances.stream()
-                    .filter(gi -> asg.getInstances().stream().noneMatch(inst -> inst.getInstanceId().equals(gi.getInstanceId()))).collect(Collectors.toList());
-            if (!asgOnlyInstances.isEmpty() || !cbOnlyInstances.isEmpty()) {
-                String errorMessage = "The instances in the autoscaling group are not in sync with the instances in cloudbreak! Cloudbreak only instances: ["
-                        + cbOnlyInstances.stream().map(CloudInstance::getInstanceId).collect(Collectors.joining(",")) + "], AWS only instances: ["
-                        + asgOnlyInstances.stream().map(Instance::getInstanceId).collect(Collectors.joining(",")) + "]. Upscale is not possible!";
-                LOGGER.info(errorMessage);
-                throw new CloudConnectorException(errorMessage);
-            }
-            if (groupInstances.size() != asg.getDesiredCapacity()) {
-                String errorMessage = String.format("The autoscale group's desired instance count is not match with the instance count in the cloudbreak."
-                        + " Desired count: %d <> cb instance count: %d. Upscale is not possible!", asg.getDesiredCapacity(), groupInstances.size());
-                LOGGER.info(errorMessage);
-                throw new CloudConnectorException(errorMessage);
-            }
-        }
-    }
 }
