+++ cloud-aws-common/src/main/java/com/sequenceiq/cloudbreak/cloud/aws/common/view/AwsNetworkView.java
@@ -1,4 +1,4 @@
-package com.sequenceiq.cloudbreak.cloud.aws.view;
+package com.sequenceiq.cloudbreak.cloud.aws.common.view;
 
 import static com.sequenceiq.cloudbreak.common.network.NetworkConstants.ENDPOINT_GATEWAY_SUBNET_ID;
 import static com.sequenceiq.cloudbreak.common.network.NetworkConstants.INTERNET_GATEWAY_ID;
