+++ cloud-aws-native/src/main/java/com/sequenceiq/cloudbreak/cloud/aws/AwsNativeSetup.java
@@ -1,0 +1,16 @@
+package com.sequenceiq.cloudbreak.cloud.aws;
+
+import org.springframework.stereotype.Component;
+
+import com.sequenceiq.cloudbreak.cloud.aws.common.AwsSetup;
+import com.sequenceiq.cloudbreak.cloud.context.AuthenticatedContext;
+import com.sequenceiq.cloudbreak.cloud.model.CloudStack;
+
+@Component
+public class AwsNativeSetup extends AwsSetup {
+
+    @Override
+    public void scalingPrerequisites(AuthenticatedContext authenticatedContext, CloudStack stack, boolean upscale) {
+
+    }
+}
